reduced files in openai
