--- python/packages/main/agent_framework/const.py
@@ -1,5 +1,0 @@
-# Copyright (c) Microsoft. All rights reserved.
-
-from typing import Final
-
-USER_AGENT: Final[str] = "User-Agent"
