+++ python/packages/main/agent_framework/openai/_shared.py
@@ -2,8 +2,10 @@
 
 import logging
 from abc import ABC
+from collections.abc import Mapping
+from copy import copy
 from enum import Enum
-from typing import Annotated, Any, Union
+from typing import Annotated, Any, ClassVar, Union
 
 from openai import (
     AsyncOpenAI,
@@ -16,13 +18,14 @@ from openai.types import Completion
 from openai.types.audio import Transcription
 from openai.types.chat import ChatCompletion, ChatCompletionChunk
 from openai.types.images_response import ImagesResponse
-from pydantic import BaseModel
+from pydantic import BaseModel, ConfigDict, Field, SecretStr, validate_call
 from pydantic.types import StringConstraints
 
 from .._logging import get_logger
-from .._pydantic import AFBaseModel
+from .._pydantic import AFBaseModel, AFBaseSettings
 from .._types import ChatOptions, SpeechToTextOptions, TextToSpeechOptions
-from ..exceptions import ServiceInvalidRequestError, ServiceResponseException
+from ..exceptions import ServiceInitializationError, ServiceInvalidRequestError, ServiceResponseException
+from ..telemetry import APP_INFO, USER_AGENT_KEY, prepend_agent_framework_to_user_agent
 from .exceptions import OpenAIContentFilterException
 
 logger: logging.Logger = get_logger("agent_framework.openai")
@@ -44,6 +47,53 @@ OPTION_TYPE = Union[
     SpeechToTextOptions,
     TextToSpeechOptions,
 ]
+
+
+class OpenAISettings(AFBaseSettings):
+    """OpenAI model settings.
+
+    The settings are first loaded from environment variables with the prefix 'OPENAI_'.
+    If the environment variables are not found, the settings can be loaded from a .env file with the
+    encoding 'utf-8'. If the settings are not found in the .env file, the settings are ignored;
+    however, validation will fail alerting that the settings are missing.
+
+    Optional settings for prefix 'OPENAI_' are:
+    - api_key: SecretStr - OpenAI API key, see https://platform.openai.com/account/api-keys
+        (Env var OPENAI_API_KEY)
+    - org_id: str | None - This is usually optional unless your account belongs to multiple organizations.
+        (Env var OPENAI_ORG_ID)
+    - chat_model_id: str | None - The OpenAI chat model ID to use, for example, gpt-3.5-turbo or gpt-4.
+        (Env var OPENAI_CHAT_MODEL_ID)
+    - responses_model_id: str | None - The OpenAI responses model ID to use, for example, gpt-4o or o1.
+        (Env var OPENAI_RESPONSES_MODEL_ID)
+    - text_model_id: str | None - The OpenAI text model ID to use, for example, gpt-3.5-turbo-instruct.
+        (Env var OPENAI_TEXT_MODEL_ID)
+    - embedding_model_id: str | None - The OpenAI embedding model ID to use, for example, text-embedding-ada-002.
+        (Env var OPENAI_EMBEDDING_MODEL_ID)
+    - text_to_image_model_id: str | None - The OpenAI text to image model ID to use, for example, dall-e-3.
+        (Env var OPENAI_TEXT_TO_IMAGE_MODEL_ID)
+    - audio_to_text_model_id: str | None - The OpenAI audio to text model ID to use, for example, whisper-1.
+        (Env var OPENAI_AUDIO_TO_TEXT_MODEL_ID)
+    - text_to_audio_model_id: str | None - The OpenAI text to audio model ID to use, for example, jukebox-1.
+        (Env var OPENAI_TEXT_TO_AUDIO_MODEL_ID)
+    - realtime_model_id: str | None - The OpenAI realtime model ID to use,
+    for example, gpt-4o-realtime-preview-2024-12-17.
+        (Env var OPENAI_REALTIME_MODEL_ID)
+    - env_file_path: str | None - if provided, the .env settings are read from this file path location
+    """
+
+    env_prefix: ClassVar[str] = "OPENAI_"
+
+    api_key: SecretStr | None = None
+    org_id: str | None = None
+    chat_model_id: str | None = None
+    responses_model_id: str | None = None
+    text_model_id: str | None = None
+    embedding_model_id: str | None = None
+    text_to_image_model_id: str | None = None
+    audio_to_text_model_id: str | None = None
+    text_to_audio_model_id: str | None = None
+    realtime_model_id: str | None = None
 
 
 class OpenAIModelTypes(Enum):
@@ -158,3 +208,87 @@ class OpenAIHandler(AFBaseModel, ABC):
             and issubclass(chat_options.response_format, BaseModel)
         ):
             options_dict["response_format"] = type_to_response_format_param(chat_options.response_format)
+
+
+class OpenAIConfigBase(OpenAIHandler):
+    """Internal class for configuring a connection to an OpenAI service."""
+
+    @validate_call(config=ConfigDict(arbitrary_types_allowed=True))
+    def __init__(
+        self,
+        ai_model_id: str = Field(min_length=1),
+        api_key: str | None = Field(min_length=1),
+        ai_model_type: OpenAIModelTypes | None = OpenAIModelTypes.CHAT,
+        org_id: str | None = None,
+        default_headers: Mapping[str, str] | None = None,
+        client: AsyncOpenAI | None = None,
+        instruction_role: str | None = None,
+        **kwargs: Any,
+    ) -> None:
+        """Initialize a client for OpenAI services.
+
+        This constructor sets up a client to interact with OpenAI's API, allowing for
+        different types of AI model interactions, like chat or text completion.
+
+        Args:
+            ai_model_id (str): OpenAI model identifier. Must be non-empty.
+                Default to a preset value.
+            api_key (str): OpenAI API key for authentication.
+                Must be non-empty. (Optional)
+            ai_model_type (OpenAIModelTypes): The type of OpenAI
+                model to interact with. Defaults to CHAT.
+            org_id (str): OpenAI organization ID. This is optional
+                unless the account belongs to multiple organizations.
+            default_headers (Mapping[str, str]): Default headers
+                for HTTP requests. (Optional)
+            client (AsyncOpenAI): An existing OpenAI client, optional.
+            instruction_role (str): The role to use for 'instruction'
+                messages, for example, summarization prompts could use `developer` or `system`. (Optional)
+            kwargs: Additional keyword arguments.
+
+        """
+        # Merge APP_INFO into the headers if it exists
+        merged_headers = dict(copy(default_headers)) if default_headers else {}
+        if APP_INFO:
+            merged_headers.update(APP_INFO)
+            merged_headers = prepend_agent_framework_to_user_agent(merged_headers)
+
+        if not client:
+            if not api_key:
+                raise ServiceInitializationError("Please provide an api_key")
+            client = AsyncOpenAI(
+                api_key=api_key,
+                organization=org_id,
+                default_headers=merged_headers,
+            )
+        args = {
+            "ai_model_id": ai_model_id,
+            "client": client,
+            "ai_model_type": ai_model_type,
+        }
+        if instruction_role:
+            args["instruction_role"] = instruction_role
+        super().__init__(**args, **kwargs)
+
+    def to_dict(self) -> dict[str, Any]:
+        """Create a dict of the service settings."""
+        client_settings = {
+            "api_key": self.client.api_key,
+            "default_headers": {k: v for k, v in self.client.default_headers.items() if k != USER_AGENT_KEY},
+        }
+        if self.client.organization:
+            client_settings["org_id"] = self.client.organization
+        base = self.model_dump(
+            exclude={
+                "prompt_tokens",
+                "completion_tokens",
+                "total_tokens",
+                "api_type",
+                "ai_model_type",
+                "client",
+            },
+            by_alias=True,
+            exclude_none=True,
+        )
+        base.update(client_settings)
+        return base
